requirements.txt
streamlit
openai
google-generativeai>=0.7.0
networkx
matplotlib
pandas
pyyaml
oci
langchain
langchain-openai
langchain-google-genai
langchain-groq
langgraph
jsonpatch
jsonpointer
numpy
requests
python-dotenv
xlsxwriter
uvicorn
fastapi
# Enhanced RAG Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
nltk>=3.8.0